Feat strip story v2 (#15)
* feat-strip-story-v2:
  - update APEStripService API.
  - fixes and updates

* cleanup

* - APEStripService
  - API update
- Demo App implementation.

* Strip Urls updates
properties fix protection levels
fixes and updates
version update

* cleanup

* - handle external links clicks
- added a spinner view
- fixes and updates

* - fix the containerView updated height
- cleanup

* fixes and updates

* revert formSubmitted, .formResubmitted handling

* - update displayStripComponent API
- fix url presentation
- remove dictionary +
