+++ Apester/ViewControllers/Strip/APEStripViewController.swift
@@ -12,54 +12,77 @@ import ApesterKit
 
 class APEStripViewController: UIViewController {
 
-  private var storyViewController: APEStripStoryViewController?
-  private let stripServiceInstance = APEStripService(channelToken: "5890a541a9133e0e000e31aa", bundle:  Bundle.main)
-  private var stripWebView: WKWebView?
+    @IBOutlet weak var containerView1: UIView!
 
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    self.stripServiceInstance.dataSource = self
-    self.stripServiceInstance.delegate = self
-    setupStripComponent()
-    // display loading view
-  }
-}
+    private var stripService: APEStripService!
 
-private extension APEStripViewController {
-  func setupStripComponent() {
-    self.stripWebView = self.stripServiceInstance.stripWebView
-    stripWebView!.frame = self.view.bounds
-    self.view.addSubview(stripWebView!)
-  }
-}
+    override func viewDidLoad() {
+        super.viewDidLoad()
 
-extension APEStripViewController: APEStripServiceDataSource {
-  var showStoryFunction: String {
-    return "console.log('show story');"
-  }
+        // set strip params
+        let params = APEStripParams(channelToken: "5890a541a9133e0e000e31aa", shape: .square, size: .medium, shadow: false, bundle: Bundle.main)
 
-  var hideStoryFunction: String {
-    return "console.log('hdie story');"
-  }
-}
-
-extension APEStripViewController: APEStripServiceDelegate {
-  func stripComponentIsReady(unitHeight height: CGFloat) {
-    self.stripWebView?.frame.size.height = height
-    /// hide loading
-    print(#function)
-  }
-
-  func displayStoryComponent() {
-    if self.storyViewController == nil {
-      self.storyViewController = APEStripStoryViewController()
-      self.storyViewController?.webView = self.stripServiceInstance.storyWebView
+        // create the StripService Instance
+        stripService = APEStripService(params: params)
+        // display the Strip Component
+        stripService.displayStripComponent(in: self.containerView1, containerViewConroller: self)
     }
-
-    self.navigationController?.pushViewController(self.storyViewController!, animated: true)
-  }
-
-  func hideStoryComponent() {
-    self.storyViewController?.navigationController?.popViewController(animated: true)
-  }
 }
+
+//class APEStripViewControllerOld: UIViewController {
+//
+//    private var storyViewController: APEStripStoryViewController?
+//    private let stripServiceInstance = APEStripService(channelToken: "5890a541a9133e0e000e31aa", bundle:  Bundle.main)
+//
+//    @IBOutlet weak var containerView: UIView!
+//
+//    //    private var stripWebView: WKWebView?
+//
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//        setupStripComponent()
+//        // display loading view
+//    }
+//}
+
+//private extension APEStripViewControllerOld {
+
+//    func setupStripComponent() {
+//        self.stripServiceInstance.dataSource = self
+//        self.stripServiceInstance.delegate = self
+//        self.stripWebView = self.stripServiceInstance.stripWebView
+//        stripWebView!.frame = self.view.bounds
+//        self.view.addSubview(stripWebView!)
+//    }
+//}
+
+//extension APEStripViewControllerOld: APEStripServiceDataSource {
+//    var showStoryFunction: String {
+//        return "console.log('show story');"
+//    }
+//
+//    var hideStoryFunction: String {
+//        return "console.log('hdie story');"
+//    }
+//}
+//
+//extension APEStripViewControllerOld: APEStripServiceDelegate {
+//    func stripComponentIsReady(unitHeight height: CGFloat) {
+//        self.stripWebView?.frame.size.height = height
+//        /// hide loading
+//        print(#function)
+//    }
+//
+//    func displayStoryComponent() {
+//        if self.storyViewController == nil {
+//            self.storyViewController = APEStripStoryViewController()
+//            self.storyViewController?.webView = self.stripServiceInstance.storyWebView
+//        }
+//
+//        self.navigationController?.pushViewController(self.storyViewController!, animated: true)
+//    }
+//
+//    func hideStoryComponent() {
+//        self.storyViewController?.navigationController?.popViewController(animated: true)
+//    }
+//}
