+++ Sources/Helpers/APEResult.swift
@@ -1,5 +1,5 @@
 //
-//  APE.swift
+//  APEResult.swift
 //  ApesterKit
 //
 //  Created by Hasan Sa on 13/07/2017.
