+++ Sources/Helpers/Extensions.swift
@@ -1,0 +1,34 @@
+//
+//  Extensions.swift
+//  ApesterKit
+//
+//  Created by Hasan Sa on 10/07/2019.
+//  Copyright © 2019 Apester. All rights reserved.
+//
+
+import CoreGraphics
+import Foundation
+
+// MARK:- String
+ extension String {
+    var dictionary: [String: Any]? {
+        if let data = self.data(using: .utf8) {
+            do {
+                return try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
+            } catch {
+                print(error.localizedDescription)
+            }
+        }
+        return nil
+    }
+}
+
+// MARK:- Float
+extension Float {
+    init?(string: String) {
+        guard let number = NumberFormatter().number(from: string) else {
+            return nil
+        }
+        self.init(number.floatValue)
+    }
+}
